tools: Adapt command line client to new output format
Adapt to the new output format that contains the sensor name.
In addition add two options --graphite-group and --graphite-group-rf
that allow to specify an optional prefix for graphite. This can be
used for example to specify the room in which the sensor is located.
As RF receivers might receive data for a different room their group
can be specified separately (sensor name: 'rf433').