main: Send Hideki sensor data from main loop
The UART transmission is interrupt based and contains a busy wait when
the receive buffer gets full. This can cause a deadlock when sending
data within an ISR where interrupts are globally disabled.

Therefore change the main application to send the Hideki Sensor data
from the main loop instead of the TimerObserver (which is called from
an ISR).